+++ extensions/vscode-api-tests/src/singlefolder-tests/configuration.test.ts
@@ -6,8 +6,11 @@
 import 'mocha';
 import * as assert from 'assert';
 import * as vscode from 'vscode';
+import { assertNoRpc } from '../utils';
 
 suite('vscode API - configuration', () => {
+
+	teardown(assertNoRpc);
 
 	test('configurations, language defaults', function () {
 		const defaultLanguageSettings = vscode.workspace.getConfiguration().get('[abcLang]');
