+++ extensions/vscode-api-tests/src/singlefolder-tests/types.test.ts
@@ -6,8 +6,11 @@
 import 'mocha';
 import * as assert from 'assert';
 import * as vscode from 'vscode';
+import { assertNoRpc } from '../utils';
 
 suite('vscode API - types', () => {
+
+	teardown(assertNoRpc);
 
 	test('static properties, es5 compat class', function () {
 		assert.ok(vscode.ThemeIcon.File instanceof vscode.ThemeIcon);
