+++ extensions/vscode-api-tests/src/singlefolder-tests/workspace.event.test.ts
@@ -5,16 +5,15 @@
 
 import * as assert from 'assert';
 import * as vscode from 'vscode';
-import { createRandomFile, withLogDisabled } from '../utils';
+import { assertNoRpc, createRandomFile, disposeAll, withLogDisabled } from '../utils';
 
 suite('vscode API - workspace events', () => {
 
 	const disposables: vscode.Disposable[] = [];
 
 	teardown(() => {
-		for (const dispo of disposables) {
-			dispo.dispose();
-		}
+		assertNoRpc();
+		disposeAll(disposables);
 		disposables.length = 0;
 	});
